[JENKINS-36501] Move RunSelector and RunFilter as top level items
