+++ src/main/java/org/jenkinsci/plugins/runselector/filters/ParameterizedRunFilter.java
@@ -32,6 +32,7 @@ import hudson.model.Run;
 import hudson.util.XStream2;
 import jenkins.model.Jenkins;
 import org.apache.commons.lang.StringUtils;
+import org.jenkinsci.plugins.runselector.RunFilter;
 import org.jenkinsci.plugins.runselector.context.RunSelectorPickContext;
 import org.kohsuke.stapler.DataBoundConstructor;
