+++ src/main/java/org/jenkinsci/plugins/runselector/filters/ParametersRunFilter.java
@@ -31,6 +31,7 @@ import hudson.model.ParametersAction;
 import hudson.model.Run;
 import hudson.model.StringParameterValue;
 import hudson.model.TaskListener;
+import org.jenkinsci.plugins.runselector.RunFilter;
 import org.jenkinsci.plugins.runselector.context.RunSelectorPickContext;
 import org.kohsuke.stapler.DataBoundConstructor;
