+++ src/main/java/org/jenkinsci/plugins/runselector/filters/RunFilterParameter.java
@@ -32,6 +32,7 @@ import hudson.model.ParameterValue;
 import hudson.model.SimpleParameterDefinition;
 import hudson.model.StringParameterValue;
 import net.sf.json.JSONObject;
+import org.jenkinsci.plugins.runselector.RunFilter;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.StaplerRequest;
