+++ src/main/java/org/jenkinsci/plugins/runselector/selectors/RunSelectorParameter.java
@@ -36,6 +36,7 @@ import hudson.model.StringParameterValue;
 import hudson.util.XStream2;
 import jenkins.model.Jenkins;
 import net.sf.json.JSONObject;
+import org.jenkinsci.plugins.runselector.RunSelector;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.StaplerRequest;
