+++ src/main/java/org/jenkinsci/plugins/runselector/selectors/SimpleRunSelectorDescriptor.java
@@ -23,6 +23,7 @@
  */
 package org.jenkinsci.plugins.runselector.selectors;
 
+import org.jenkinsci.plugins.runselector.RunSelector;
 import org.jvnet.localizer.Localizable;
 
 /**
